Implement TimerEntity and saving to database

+++ subathon-dataservice/src/main/java/com/pronixxx/subathon/data/entity/BaseEntity.java
@@ -16,16 +16,16 @@ public abstract class BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    protected Integer id;
+    protected long id;
 
     @Column(name = "insert_time")
     protected Instant insertTime;
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
